+++ CONTRIBUTING.md
@@ -1,3 +1,3 @@
-Contributing to fast_barcode_scanner
+Contributing to icapps_fast_barcode_scanner
 
 Contributions are welcome by submitting a PR for to be reviewed.
