+++ fast_barcode_scanner/android/build.gradle
@@ -1,4 +1,4 @@
-group 'com.jhoogstraat.fast_barcode_scanner'
+group 'com.jhoogstraat.icapps_fast_barcode_scanner'
 version '1.0-SNAPSHOT'
 
 buildscript {
