+++ fast_barcode_scanner/android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'fast_barcode_scanner'
+rootProject.name = 'icapps_fast_barcode_scanner'
