+++ fast_barcode_scanner/android/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.jhoogstraat.fast_barcode_scanner">
+  package="com.jhoogstraat.icapps_fast_barcode_scanner">
     <uses-feature android:name="android.hardware.camera.any" />
     <uses-permission android:name="android.permission.CAMERA"/>
 </manifest>
