+++ fast_barcode_scanner/android/src/main/kotlin/com/jhoogstraat/icapps_fast_barcode_scanner/Types.kt
@@ -1,4 +1,4 @@
-package com.jhoogstraat.fast_barcode_scanner
+package com.jhoogstraat.icapps_fast_barcode_scanner
 
 import android.util.Size
 import com.google.mlkit.vision.barcode.Barcode
