+++ fast_barcode_scanner/example/android/app/src/debug/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.jhoogstraat.fast_barcode_scanner_example">
+    package="com.jhoogstraat.icapps_fast_barcode_scanner_example">
     <!-- Flutter needs it to communicate with the running application
          to allow setting breakpoints, to provide hot reload, etc.
     -->
