+++ fast_barcode_scanner/example/android/app/src/main/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.jhoogstraat.fast_barcode_scanner_example">
+    package="com.jhoogstraat.icapps_fast_barcode_scanner_example">
    <application
-        android:label="fast_barcode_scanner_example"
+        android:label="icapps_fast_barcode_scanner_example"
         android:icon="@mipmap/ic_launcher">
         <activity
             android:name=".MainActivity"
