+++ fast_barcode_scanner/example/android/app/src/main/kotlin/com/jhoogstraat/fast_barcode_scanner_example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.jhoogstraat.fast_barcode_scanner_example
+package com.jhoogstraat.icapps_fast_barcode_scanner_example
 
 import io.flutter.embedding.android.FlutterActivity
