+++ fast_barcode_scanner/example/pubspec.yaml
@@ -10,14 +10,14 @@ dependencies:
   flutter:
     sdk: flutter
 
-  fast_barcode_scanner:
+  icapps_fast_barcode_scanner:
     path: ../
 
 dev_dependencies:
   flutter_lints: ^1.0.4
 
 dependency_overrides: 
-  fast_barcode_scanner_platform_interface:
+  icapps_fast_barcode_scanner_platform_interface:
     path: ../../fast_barcode_scanner_platform_interface/
 
 flutter:
