--- fast_barcode_scanner/lib/fast_barcode_scanner.dart
@@ -1,13 +1,0 @@
-export 'package:fast_barcode_scanner/src/barcode_camera.dart';
-export 'package:fast_barcode_scanner/src/camera_controller.dart';
-export 'package:fast_barcode_scanner/src/overlays/blur_overlay.dart';
-export 'package:fast_barcode_scanner/src/overlays/material_overlay.dart';
-export 'package:fast_barcode_scanner_platform_interface/fast_barcode_scanner_platform_interface.dart'
-    show
-        Barcode,
-        BarcodeType,
-        Framerate,
-        Resolution,
-        DetectionMode,
-        CameraPosition,
-        PreviewConfiguration;
