+++ fast_barcode_scanner/lib/icapps_fast_barcode_scanner.dart
@@ -1,0 +1,13 @@
+export 'package:icapps_fast_barcode_scanner/src/barcode_camera.dart';
+export 'package:icapps_fast_barcode_scanner/src/camera_controller.dart';
+export 'package:icapps_fast_barcode_scanner/src/overlays/blur_overlay.dart';
+export 'package:icapps_fast_barcode_scanner/src/overlays/material_overlay.dart';
+export 'package:icapps_fast_barcode_scanner_platform_interface/icapps_fast_barcode_scanner_platform_interface.dart'
+    show
+        Barcode,
+        BarcodeType,
+        Framerate,
+        Resolution,
+        DetectionMode,
+        CameraPosition,
+        PreviewConfiguration;
