+++ fast_barcode_scanner/lib/src/camera_controller.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
-import 'package:fast_barcode_scanner_platform_interface/fast_barcode_scanner_platform_interface.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:icapps_fast_barcode_scanner_platform_interface/icapps_fast_barcode_scanner_platform_interface.dart';
 
 class CameraConfiguration {
   const CameraConfiguration(
