+++ fast_barcode_scanner/pubspec.yaml
@@ -1,6 +1,6 @@
-name: fast_barcode_scanner
+name: icapps_fast_barcode_scanner
 description: A fast barcode scanner using MLKit on Android and AVFoundation on iOS.
-version: 1.1.4
+version: 2.1.0
 homepage: https://github.com/jhoogstraat/fast_barcode_scanner
 repository: https://github.com/jhoogstraat/fast_barcode_scanner
 
@@ -10,7 +10,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  fast_barcode_scanner_platform_interface: ^1.0.4
+  icapps_fast_barcode_scanner_platform_interface: ^1.2.0
 
 dev_dependencies:
   flutter_test:
@@ -21,7 +21,7 @@ flutter:
   plugin:
     platforms:
       android:
-        package: com.jhoogstraat.fast_barcode_scanner
+        package: com.jhoogstraat.icapps_fast_barcode_scanner
         pluginClass: FastBarcodeScannerPlugin
       ios:
         pluginClass: FastBarcodeScannerPlugin
