+++ fast_barcode_scanner/test/fast_barcode_scanner_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  const MethodChannel channel = MethodChannel('fast_barcode_scanner');
+  const MethodChannel channel = MethodChannel('icapps_fast_barcode_scanner');
 
   TestWidgetsFlutterBinding.ensureInitialized();
