--- fast_barcode_scanner_platform_interface/lib/fast_barcode_scanner_platform_interface.dart
@@ -1,4 +1,0 @@
-export 'package:fast_barcode_scanner_platform_interface/src/fast_barcode_scanner_platform_interface.dart';
-export 'package:fast_barcode_scanner_platform_interface/src/types/barcode.dart';
-export 'package:fast_barcode_scanner_platform_interface/src/types/barcode_type.dart';
-export 'package:fast_barcode_scanner_platform_interface/src/types/preview_configuration.dart';
