+++ fast_barcode_scanner_platform_interface/lib/icapps_fast_barcode_scanner_platform_interface.dart
@@ -1,0 +1,4 @@
+export 'package:icapps_fast_barcode_scanner_platform_interface/src/fast_barcode_scanner_platform_interface.dart';
+export 'package:icapps_fast_barcode_scanner_platform_interface/src/types/barcode.dart';
+export 'package:icapps_fast_barcode_scanner_platform_interface/src/types/barcode_type.dart';
+export 'package:icapps_fast_barcode_scanner_platform_interface/src/types/preview_configuration.dart';
