+++ fast_barcode_scanner_platform_interface/pubspec.yaml
@@ -1,8 +1,8 @@
-name: fast_barcode_scanner_platform_interface
+name: icapps_fast_barcode_scanner_platform_interface
 description: A common platform interface for the fast_barcode_scanner plugin.
 homepage: https://github.com/jhoogstraat/fast_barcode_scanner/tree/master/fast_barcode_scanner_platform_interface
 repository: https://github.com/jhoogstraat/fast_barcode_scanner
-version: 1.0.4
+version: 2.2.0
 
 environment:
   sdk: '>=3.0.0 <4.0.0'
